Update Configure new project screen
When you set the application name to "ComponentCase" and Company domain to "phonegapday.com" it automatically sets the Package name to "com.phonegapday.componentcase". If we don't edit the package name at this step the PGDayEU16Plugin class will fail to build as it won't have an import for "MainActivity". Even when you add the import the getBookmarks and addItem methods are protected and since the classes are in different packages they won't be available.

I'm just suggesting we make this an explicit instruction here instead of running into problems later in the tutorial.